Added a firing bullet mechanic
In a certain state, the player can now fire a bullet, which appears on the player position, is oriented towards the position of the click, and follows that direction. I had to mess with the collision layer and mask to prevent the bullet from colliding with the player, I'm not sure what configuration is better, but it looks like a pain to do.