Bug fix
Fixed a bug that caused non-fatal errors on load. The on_signal_health_changed function of the health_changed_manager.gd script was trying to add_child to a node that was busy, so it couldn't. An easy fix was to defer that, but then it showed heals at the start, so I searched for the real problem, which was that, in health_component.gd, the health was initialized in _ready() so it used the setter of health. Now the health is initialized at variable declaration. So, fixed.