Bug fix + improvement + cleaning
Fixed the bug where the rocket would not deal damage in AoE if the player was in range. Fixed it by moving the damage part inside the if statement testing if it is a player.

After that, improved that part of the script by making it so that it can be chosen whether or not the rocket deals self damage or not, it still has to be linked to the options but this is a good start.

Also improved clarity of the code a bit.