Bug issue resolved + cleaning
Resolved the bug where the damage label would not show on the killing blow, but adding a timer before the queue free so that the label has time (LoL) to be seen.

Added a new CharacterClass to handle the death in a grouped way.

Created a new issue where the characters can still be affected by the environment (eg damaged) after death.

Cleaned some piece of the code.