Solved both the previous issues
Fixed the issue of the AoE of bullet 3 (that I caused myself).

Finally fixed the issue of the label on killing damage. Plus I made it look cool with the superclass (that now contains the deactivate_node() function).

Set new objective of handling better the death of the characters, by adding animation, sound and handle the part where they exit the screen and disappear.