Fixed an issue
Fixed the issue where the dead bodies would not fall when they were killed by melee weapon. The issue was an error caused because the deactivate call couldn't modify the thing because of flushing queries (cf Notes.md, I don't really understand it LoL).

Now I can focus on the disappearing of the bodies when out of screen.