Added destruction of dead out-of-screen enemies
Now, when a character dies, it falls through the map, and when it exits the screen, after a fixed amount of time, it disappears (queue_free()). This does not yet apply to the player, though it should not be a problem to do it (the code is already there), because I'm not sure if it is a good idea, because in the end, there should be stuff happening when the player dies, like a respawn or something, so it is useless for now to allow efforts to something that will probably be reverted later.