Changed a few parameters + modified the map + added functional movement for the enemies
Changed the parameters of display, from a 960 * 540 Viewport to 1920 * 1080, from windowed to fullscreen, stretch mode canvas_item and aspect expand, and from canvas texture filter linear to nearest. Basically, the window is bigger, the pixel art is rendered better, and the game is not too zoomed in, but maybe this is a mistake, I'm not too sure. Anyway, this can always be changed later (it might be necessary, for compatibility with smaller and bigger screens.
Actually this is a bad idea for now, I need the small window to debug.

Modified the map so that the enemies don't fall on their own.

Added the code necessary to the movement of the enemies, for now they just wander left, idle, and right, but it is with random values, so it is cool. Of course, they don't know how to react to walls yet, but it is a good start.