Improved health + few changes
Improved how the health: max_health is now a character superclass variable. The setter now checks if the health set value doesn't exceed max_health.

Commented a bit.

The move_speed variable of character is now the speed of follow (soon to be), and wander_speed is in [move_speed/4, move_speed/2].