Full implementation of AttackState (except hit transition) + other changes
That's right, fully implemented the AttackState, that means on enter code, on exit code, code to transition to FollowState (not to HitState though), attack cooldown, not attack itself but almost, with correct cooldown.

Completed code in superclasses and main enemy script. Moved CMSM.

Also commented a bit, and added static type to a lot of routines. Made some stuff clearer.