Solved a bug in HitState
There was a bug in HitState: when the enemy got hit twice in a short amount of time (before the end of the "stun animation"), it reacted wrong to future hits when in WanderState, taht is to say, it was only getting stunned (ie in the HitState) for a split second. This has been resolved, after hours of arduous work, by changing the way that the transition to WanderState was made (from using next_state to using interrupt_state()).