Fixed AttackState
The enemy used to attack only when entering the zone, not when staying in it. now it does (I did the same thing I did to the player counterpart, that is to say setting the monitoring of the area back to false just after a timer, right after setting it to true).