Changed Z index again
The characters were not correctly displayed behind the grass as they should be (only the player was, but it was because of the order of the scenes in the level scene). I changed the Z index of the characters (one by one, don't know a better way). This broke the previous change on the labels, so I fixed it as well, now it's all good.

Also, uncovered a weird thing: the attack of the enemy seems to come only a while after the player entered the range. This is the objective, but it shouldn't be the case right now, so it is weird.