Fixed a bug with Bullet 3 AoE and TME
Fixed a bug of the AoE of Bullet 3 interacting incorrectly with the areas of the TME (it had two areas that were detected by the explosion, so it was damaged twice per shot). Now it is fixed, the AoE checks if the area is of the correct type before damaging its parent, should not cause any more problems.

Also, tweaked a few things.