Broke the bullet 3 AoE and damage
It is broken for now, but there still is the old code so it is an easy fix. I don't think the code needs much more to function.
Let me explain: I tried to make the AoE damage depend on the distance from the explosion, but it was difficult or impossible by using overlapping areas, so I tried ray-casting. It is not working as it should yet, but it is on the right tracks. However, it might not be the expected result, as the ray, thus the damage, stops at the first encounter. This might work for a different weapon, possibility to be explored.
Anyway, I'm pushing to keep track.