Lot of bullshit, and new "matrix" damage technique
Title says it all. A lot of stuff has been changed, I created a new bullet_3.gd script to get rid of the old stuff without getting rid of it, part of that stuff being a new technique tried to make damages in a nova (expanding area) that led to nothing.
I am tired so I am commiting before I lose it.