Not working expanding AoE and working simple AoE
I tried to make the AoE expand (several overlapping areas), but for some reason it doesn't work. I'll have to figure this out tomorrow.
I re-added the code and nodes for the "simple" AoE, that is to say just an area that does damage (it doesn't actually do damage, but it accurately detects other areas, so that's a win in my book). This works, so at least, after all this shit, I can go back to the basic stuff that actually works.