Fixed the AoE bug
Fixed the bug in bullet_3.gd where the enemy wouldn't take full damage, depending on the size of the AoE. It was caused by the fact that the areas where not correctly positioned: they were at the center of the missile, when they should be at the tip. Fixed now.