Fixed damage label + minor changes
Fixed the issue with the damage label, that would not display the damages correctly if they were several on the same frame. Now it does (there still is an issue if heal and damage are on the same frame, but it is ok for now).

Commented a bit

Changed the code for bullet 3 from overlapping areas to bodies.