Moved and renamed stuff
Renamed the weapon reload variables and made them export for easier debugging.

Moved code for weapon attacks from state_process() to state_input(), because it is better suited for that kind of work, and also because I was trying to understand how it works. The event of state_event() happens simply when a new input is given, not when a key is continuously pressed.