Solved issue of Z-index when dying
When a character dies, it falls through the ground. But, because the character has to be behind the grass, it is also behind the ground (else, it would require a second tile map or something; actually, super easy, barely an inconvenience).
I was going for changing the Z-order of the characters at runtime when they die, but it is easier to change that of the tile map, as it can be set for each individual tile, therefore can be low for the tile map, and high for the grass tiles. Changed the Z-index of the player, don't know why it was at -1, might cause issues in the future. Actually, it is weird that the character falls behind the grass but in front of the ground...
Should also work for the labels.
Actually, went for the first solution because otherwise there are issues, cf Notes.md.