Almost finished Fire Spitter
The code for Fire Spitter is mostly done, that is to say that there is an area in front of the player, that rotates to follow the mouse, with a semi-transparent sprite that flips when on the other side attached to it, and that is hidden when not in use. When Fire Spitter is the main weapon, nothing happens if no input. On click, the sprite appears, the hit box activates, damaging the enemies inside, then deactivates. On click hold, the enemies inside are damaged continuously.
What this lacks, it the setting on fire part, the visuals (animations) and the balancing.