Kinda prevented a bug
When the player is queue freed, is crashes because of the enemies following the player. I could try to go around that by queue freeing everything before the player, that might cause issues, Idk, or by having the enemies not reference the player directly, but instead look for him in a range, which would probably be expensive. So, for now at least, I just don't free the player.