Added burning ground mechanic
Added burning ground zone mechanic, which will be used for Flame Thrower. It might also be useful in the future as an obstacle mechanic, but I think I'd have to make it another scene, otherwise it would be too complicated. Maybe they could have a common superclass?

Also, added base values for variables in the functions in health_component.gd, so that it doesn't crash if I forget to add them on call.