Correctly firing Flame Thrower
The Flame Thrower now fires projectiles that are subject to gravity and air resistance, and is an automatic weapon. Now, apart from visuals and sound, there is only need to connect the ground fire and the projectile, and to make the projectiles face their movement direction, maybe. So, I added a few functions to bullet.gd, added napalm scene and code, and almost all of the Flame Thrower code (copy of Machine Gun code)