Finished the Flame Thrower
Finished the code for the Flame Thrower, except the part where the projectiles face the direction of their movement, because I don't want to do it and I think it is useless. Also, it lacks the connection of the damage from the weapon to the fire. Also, maybe it would be good to set the target on fire just by touching them, not only when they walk over the burning ground.
And ofc there are no visuals nor sounds.