Fixed stuff, finished Flame Thrower
Fixed the burning ground, it now correctly refreshes the burning.

Fixed the fire code in health component so that it also correctly refreshes the burn.

Connected the Flame Thrower variables to the actual fire variables.

Made the enemy hit burn on napalm projectile contact.

FIxed a bug where call set_on_fire() with a fire_duration of 0 would still damage once. Should not be very useful, but still.

Decided that it was ok that the burning ground spawns at the feet of the enemies and not under them, because it would be weird for a big enemy, it doesn't really make sense that the liquid would move on its own, and anyway most of the enemies will be moving so it doesn't really matter.