Finished the Grenade Launcher
Finished the code for the Grenade Launcher, that means I added the fact that the explosion of the grenade creates fragments that fly and explode on contact. Not sure yet of the speed they should have.

Also, fixed an issue where the health changed label would display a lot of decimals when the rocket aoe size was not set to a "round" value (a multiple of the damage or something like that).  Actually, I'll fix it later.

Probably something else that I forget.