Added an easter egg
Added an easter egg where, if a certain value is set correctly, which should happen is options when they come out, a sound (for now, the chorus of "Allumer le feu") is played when a character is set on fire. That raises the issue of copyrights, but for now it works and that's all that matters.