Boxed the test level
Made a box around the level, so that the projectiles do not continue flying forever, thus limiting the impact on performances. The only possible issue is that it might be in the way of having a bigger level, but that is easy to fix, just need to expand the box.