Finished weapon throw
The throwing weapon feature is now complete, with the weapon having a sprite (axe) ,rotating on hit, able to be thrown away, rotating and damaging on its path and following the mouse, then returning at high speed to the player (still rotating and damaging). The damage/rotation stops when returned to player, but also is paused when changing weapons (this hides the axe when on player), then the weapon stays in the air. It is kinda weird that its position is linked to that of the player, but it is okay. Also, maybe the flying weapon should pause mid air before returning, Idk.