Almost finished Sound Blaster + some stuff
The Sound Blaster is almost finish, it only lacks solving an issue with knocking back the moving enemies, and playing the sounds.

Also, refactored the file layout a bit.

Also, improved the allumer le feu shit, now the only possible issue is the fact that several sounds can be played at the same time.