Edited sounds for Sound Blaster + minor changes
Edited (cut the uninteresting parts) the sounds for the Sound Blaster, but forgot to normalize the sound volume. Can do that later anyway.
Also, made it so that the played sound cannot be the same as the one previously played, except if there is only one sound in the list.