Beginning of Health bar + bug fix
Started to create the basis of health bars for the characters.

Also, fixed the bug where healing for 10 on a character with max_health - 5 health (for instance) would not display a heal. Now it display a +10 heal. Could display +5, but I think it is better to display the full heal.