Functional health and reload bars
Made health bars for all the characters, just above their tag, that is filled at the beginning, and then empties as the health decreases, and fills at it increases.

Made a reload bar for the player, that shows the progress of the reload, for each weapon, depending on the reload time of the weapon.
Two issues: the throw calls hit() one too many times, causing some weird shit with the reload bar, and the reload bar doesn't work well if the weapon changes during reload (the bar resets, thus desynchronizing with the actual reload time left). Will see that tomorrow.