Partially functional blink
Made a blink code that works, inside character superclass, depending on whether or not the character can blink. It teleports the character to the chosen position if it is close by, or to the direction times the max distance otherwise. It plays a sound when blinking, using the same robust code used for the Sound Blaster (nothing is hard-coded, it can handle any amount of sounds, and plays them randomly).
What it does not do yet, is check if the position is acceptable, nor have a cool down and an indicator bar.