Almost solved blink into wall
The issue of blinking inside terrain is almost solved, now it checks if the position is inside something, and teleports only if the space is clear. Though that could stay this way, moving only if the position is valid, I am trying to have the character slide to a correct location if it is not the case.