Finished Blink
Blink is now fully functional, teleporting the character to the pointed location or the farthest one possible, checking if the teleportation is inside a wall, and not teleporting if it is the case, but instead kinda gliding there (this is not perfect, the character can get stuck on the ground and shit, but it isn't too bad). The blink functions on a cool down, and uses, if it exists, a reload bar to display the remaining cool down.