Added walkable slopes + fixed a bug
Added physics to the slopes of the tile map, made it so the transitions are smooth (the characters can walk up no problem), made it so that the characters can actually slide down slopes instead of jaggering down (increased the floor snap length to 5).

Also fixed a bug where the burning ground would not completely disappear after the timer ran out (the queue_free() was missplaced).